[HUDI-2137] Make hudi examples runnable locally
