+++ hudi-examples/hudi-examples-spark/bin/hudi-delta-streamer
