--- hudi-examples/src/main/java/org/apache/hudi/examples/common/ExampleDataSchemaProvider.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hudi.examples.common;
-
-import org.apache.avro.Schema;
-import org.apache.hudi.common.config.TypedProperties;
-import org.apache.hudi.utilities.schema.SchemaProvider;
-import org.apache.spark.api.java.JavaSparkContext;
-
-
-/**
- * the example SchemaProvider of example json data from uber.
- */
-public class ExampleDataSchemaProvider extends SchemaProvider {
-
-  public ExampleDataSchemaProvider(TypedProperties props, JavaSparkContext jssc) {
-    super(props, jssc);
-  }
-
-  @Override
-  public Schema getSourceSchema() {
-    return HoodieExampleDataGenerator.avroSchema;
-  }
-}
